fix again request for haimi
