+++ src/java/Controller/ProfileUserUpdate.java
@@ -5,6 +5,7 @@
 package Controller;
 
 import Dao.Impl.UserInformationDaoImpl;
+import Model.User;
 import Model.UserInformation;
 import java.io.IOException;
 import jakarta.servlet.ServletException;
@@ -30,12 +31,15 @@ public class ProfileUserUpdate extends HttpServlet {
     @Override
     protected void doGet(HttpServletRequest request, HttpServletResponse response)
             throws ServletException, IOException {
-        int id = Integer.parseInt(request.getParameter("id"));
-        UserInformationDaoImpl userInformationDaoImpl = new UserInformationDaoImpl();
-        UserInformation userinf = userInformationDaoImpl.get(id);
-        request.setAttribute("userinf", userinf);
-        request.getRequestDispatcher("/user/profile/ProfileUpdate.jsp").forward(request, response);
-
+        User user = (User) request.getSession().getAttribute("user");
+        if (user == null) {
+            response.sendRedirect(request.getContextPath());
+        } else {
+            UserInformationDaoImpl userInformationDaoImpl = new UserInformationDaoImpl();
+            UserInformation userinf = userInformationDaoImpl.get(user.getId());
+            request.setAttribute("userinf", userinf);
+            request.getRequestDispatcher("/user/profile/ProfileUpdate.jsp").forward(request, response);
+        }
     }
 
     /**
