+++ src/java/Dao/Impl/UserInformationDaoImpl.java
@@ -6,7 +6,7 @@ package Dao.Impl;
 
 import Dao.UserInformationDao;
 import Model.UserInformation;
-import DBContext.DBContext;
+import Dao.DBContext;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
 import java.sql.SQLException;
